Stop sampling if events are not allowed to be sent
Right now `Transaction#set_initial_sample_decision` only checks if
tracing is enabled before sampling but not if events are allowed to be
sent (mainly due to environments).

Although the sampled transactions won't really be sent because
`Client#send_event` will reject them. The SDK would still log sampling
messages, which is noisy and confusing. So this commit fixes the issue.